config: introduce config builder
